requirements.txt
opencv-python==4.8.0.76
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
tensorflow==2.13.0
# face-recognition==1.3.0  # Comment lại hoặc xóa dòng này
# dlib==19.24.1            # Comment lại hoặc xóa dòng này
h5py==3.9.0
pillow==10.0.0
mtcnn==0.1.1 